Add buffer search overloads, refactor table row/pack logic
Added overloads for `buffer_find_g` and `buffer_find_last_g` to support searching for single byte values (`uint8_t`) and other integer types, with overload resolution and documentation. Changed `table_base::offset_find_row` to return `int64_t` for consistency, updated related methods, and refactored `pack_plant_span` to use `arrayPack` and call `row_add_pack_one`. Added overloads for `pack_harvest_span`, a new `pack_extract_iter` template, and improved variable naming for Hungarian notation compliance. Updated tests for comment stripping to use new buffer search logic and improved comment boundary handling. Disabled some test cases for clarity.

+++ external/gd/gd_table_simd.cpp
@@ -807,7 +807,7 @@ int table_base::offset_find_column(uint64_t uOffset, gd::types::tag_size8) const
  * @param tag_size8 tag indicating 8-bit size
  * @return row index or -1 if not found
  */
-int table_base::offset_find_row(uint64_t uOffset, gd::types::tag_size8) const noexcept
+int64_t table_base::offset_find_row(uint64_t uOffset, gd::types::tag_size8) const noexcept
 {                                                                                                  assert(m_uRowSize > 0 && size_pack() > 0);
    const uint64_t uTotalRowDataSize = m_uRowCount * m_uRowSize;                                             
    if(uOffset >= uTotalRowDataSize) return -1;
@@ -817,7 +817,7 @@ int table_base::offset_find_row(uint64_t uOffset, gd::types::tag_size8) const no
    const uint64_t uColumn = uOffsetInRowPack / size_pack();                                        assert(uColumn < get_column_count());
    const uint64_t uRowIndex = uRow * get_column_count() + uColumn;                                 assert(uRowIndex < (m_uRowCount * get_column_count()));
 
-   return static_cast<int>(uRowIndex);
+   return static_cast<int64_t>(uRowIndex);
 }
 
 /** ---------------------------------------------------------------------------
